QCkit: NPS Inventory and Monitoring Quality Control Toolkit

This package contains a set of useful functions for data munging, quality control, and data flagging. Functions are contributed by multiple U.S. National Park Service staff, contractors, partners and others. These functions will likely be most useful for quality control of NPS data but may have utility beyond their intended functions.

See also

Author

Maintainer: Robert Baker robert_baker@nps.gov (ORCID)

Authors:

Judd Patterson judd_patterson@nps.gov (ORCID)

Joe DeVivo

Issac Quevedo (ORCID)

Sarah Wright (ORCID)

Other contributors:

Sarah Kelso (ORCID) [contributor]

Amy Sherman (ORCID) [contributor]