Coordinate Conversion from UTM to Latitude and Longitude

UTMtoLL takes your dataframe with UTM coordinates in separate Easting and Northing columns, and adds on an additional two columns with the converted decimalLatitude and decimalLongitude coordinates using the reference coordinate system WGS84.

UTMtoLL(Data, EastingCol, NorthingCol, Zone, Datum = "WGS84")

Arguments

Data

- The dataframe with UTM coordinates you would like to convert. Input the name of your dataframe.

EastingCol

- The name of your Easting UTM column. Input the name in quotations, ie. "EastingCol".

NorthingCol

- The name of your Northing UTM column. Input the name in quotations, ie. "NorthingCol".

Zone

- The UTM Zone. Input the zone number in quotations, ie. "17".

Datum

- The datum used in the coordinate reference system of your coordinates. Input in quotations, ie. "WGS84"

Value

The function returns your dataframe, mutated with an additional two columns of decimal Longitude and decimal Latitude.

Details

Define the name of your dataframe, the easting and northing columns within it, the UTM zone within which those coordinates are located, and the reference coordinate system (datum). UTM Northing and Easting columns must be in separate columns prior to running the function.

If a datum is not defined, the function will default to "WGS84".

If there are missing coordinates in your dataframe they will be preserved, however they will be moved to the end of your dataframe.

Examples

if (FALSE) {
UTMtoLL(Data = mydataframe, 
        EastingCol = "EastingCoords", 
        NorthingCol = "NorthingCoords", 
        Zone = "17", 
        Datum = "WGS84")
}